Ground up property mapping without AutoMap dependency
